fix: some small cleanups
